Moving Disease in seperated file
